requirements.txt
Django==1.6.4
Pygments==1.6
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.5
django-toolbelt==0.0.1
djangorestframework==2.3.13
fancycompleter==0.4
gunicorn==18.0
honcho==0.5.0
ipdb==0.8
ipython==2.0.0
pdbpp==0.7.2
psycopg2==2.5.2
pyrepl==0.8.4
pystache==0.5.3
static==1.0.2
wmctrl==0.1
wsgiref==0.1.2
